+++ webapp/components/channel_view.jsx
@@ -13,7 +13,6 @@ import PostsViewContainer from 'components/posts_view_container.jsx';
 import ChannelStore from 'stores/channel_store.jsx';
 
 import * as Utils from 'utils/utils.jsx';
-import * as GlobalActions from 'action_creators/global_actions.jsx';
 
 export default class ChannelView extends React.Component {
     constructor(props) {
@@ -42,23 +41,6 @@ export default class ChannelView extends React.Component {
         ChannelStore.addChangeListener(this.updateState);
 
         $('body').addClass('app__body');
-
-        document.addEventListener('keydown', (e) => {
-            if (e.altKey && (e.keyCode === Constants.KeyCodes.UP || e.keyCode === Constants.KeyCodes.DOWN)) {
-                const allChannels = ChannelStore.getAll();
-                const curChannel = this.state.channel;
-                const curIndex = allChannels.find(curChannel);
-                let nextChannel = curChannel;
-                let nextIndex = curIndex;
-                if (e.keyCode === Constants.KeyCodes.DOWN) {
-                    nextIndex = Math.min(curIndex + 1, allChannels.length - 1);
-                } else if (e.keyCode === Constants.KeyCodes.UP) {
-                    nextIndex = Math.max(curIndex - 1, 0);
-                }
-                nextChannel = allChannels[nextIndex];
-                GlobalActions.emitChannelClickEvent(nextChannel);
-            }
-        });
     }
     componentWillUnmount() {
         ChannelStore.removeChangeListener(this.updateState);
